# swagger/swaggerDEV.json
{
  "swagger": "2.0",
  "info": {
    "description": "Backend ::ASA-API-RESTFULL:: App",
    "version": "1.0.0",
    "title": "Swagger ASA-API-RESTFULL",
    "contact": {
      "email": "jcanocarballo@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "https://asa-api-restfull.herokuapp.com",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "paths": {
    "/auth": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login del usuario",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "Login",
            "description": "Objeto de Login",
            "required": true,
            "schema": {
                "$ref": "#/definitions/body_login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema":{
              "type":"object",
              "$ref":"#/definitions/Response_login" 
            }         
          },
          "400": {
            "description": "Invalid status value",
            "schema":{
              "type":"object",
              "$ref":"#/definitions/400_login"
            }            
          }
        }
      }
    },
    "/auth/register":{
      "post":{
        "tags":["Auth"],
        "summary":"Registrar usuario",
        "description":"",
        "produces":["application/json"],
        "parameters":[
          {
            "in": "body",
            "name": "Register",
            "description": "Objeto de register",
            "required": true,
            "schema": {
                "$ref": "#/definitions/body_register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema":{
              "type":"object",
              "$ref":"#/definitions/Response_login" 
            }         
          },
          "400": {
            "description": "Invalid status value"                    
          }
        }
      }
    },
    "/auth/renew": {
      "get":{
        "tags":["Auth"],
        "summary":"Renovar token",
        "description":"",
        "produces":["application/json"],
        "parameters":[
          {
            "in": "header",
            "name": "x-token",
            "description": "Token de authenticación",
            "required": true,
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDk4YWI5NzMyMTQyZTA3ZjA2NjU0YjgiLCJuYW1lIjoiSnVsaWFuIENhbm8iLCJlbWFpbCI6ImpjYW5vY0BnbWFpbC5jb20iLCJpYXQiOjE2MjA2NDg2MDEsImV4cCI6MTYyMDY1NTgwMX0.zBMjzcy04a5hJD71eI6FymdAQ8olBgU7V6uX78Mzw-o"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema":{
              "type":"object",
              "$ref":"#/definitions/Response_login" 
            }         
          },
          "400": {
            "description": "Invalid status value",
            "schema":{
              "type":"object",
              "$ref":"#/definitions/400_renew"
            }            
          }
        }
      }
    },
    "/categorias": {
      "get": {
        "tags": ["Categorias"],
        "summary": "Recupera lista de categorias",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "x-token",
            "description": "Token de authenticación",
            "required": true,
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDk4YWI5NzMyMTQyZTA3ZjA2NjU0YjgiLCJuYW1lIjoiSnVsaWFuIENhbm8iLCJlbWFpbCI6ImpjYW5vY0BnbWFpbC5jb20iLCJpYXQiOjE2MjA2NDg2MDEsImV4cCI6MTYyMDY1NTgwMX0.zBMjzcy04a5hJD71eI6FymdAQ8olBgU7V6uX78Mzw-o"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa.",
            "schema": {
              "$ref": "#/definitions/listaCategoria"
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      },
      "post":{
        "tags": ["Categorias"],
        "summary": "Registra categoria",
        "description": "",
        "produces": ["application/json"],
        "parameters": [          
          {
            "in": "header",
            "name": "x-token",
            "description": "Token de authenticación",
            "required": true,
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDk4YWI5NzMyMTQyZTA3ZjA2NjU0YjgiLCJuYW1lIjoiSnVsaWFuIENhbm8iLCJlbWFpbCI6ImpjYW5vY0BnbWFpbC5jb20iLCJpYXQiOjE2MjA2NDg2MDEsImV4cCI6MTYyMDY1NTgwMX0.zBMjzcy04a5hJD71eI6FymdAQ8olBgU7V6uX78Mzw-o"
          },
          {
            "in": "body",
            "name": "Categoria",
            "description": "Objeto de categoria",
            "required": true,
            "schema": {
                "$ref": "#/definitions/Categoria"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa.",
            "schema": {
              "type": "object",
              "properties":{
                "ok":{
                  "type":"boolean",
                  "example":true
                },
                "msg": {
                  "type":"string",
                  "example":"Registro exitoso"
                }
              }
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      }      
    },
    "/categorias/{id}":{
      "put": {
        "tags": ["Categorias"],
        "summary": "Actualizar una categoria",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Categoria Id",
            "required": true
          },
          {
            "in": "header",
            "name": "x-token",
            "description": "Token de authenticación",
            "required": true,
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDk4YWI5NzMyMTQyZTA3ZjA2NjU0YjgiLCJuYW1lIjoiSnVsaWFuIENhbm8iLCJlbWFpbCI6ImpjYW5vY0BnbWFpbC5jb20iLCJpYXQiOjE2MjA2NDg2MDEsImV4cCI6MTYyMDY1NTgwMX0.zBMjzcy04a5hJD71eI6FymdAQ8olBgU7V6uX78Mzw-o"
          },
          {
            "in": "body",
            "name": "Categoria",
            "description": "Objeto de categoria",
            "required": true,
            "schema": {
                "$ref": "#/definitions/Categoria"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties":{
                "ok":{
                  "type":"boolean",
                  "example":true
                },
                "msg": {
                  "type":"string",
                  "example":"El registro fue actualizado con exito."
                }
              }
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      },
      "delete": {
        "tags": ["Categorias"],
        "summary": "Elimina una categoria",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Categoria Id",
            "required": true
          },
          {
            "in": "header",
            "name": "x-token",
            "description": "Token de authenticación",
            "required": true,
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDk4YWI5NzMyMTQyZTA3ZjA2NjU0YjgiLCJuYW1lIjoiSnVsaWFuIENhbm8iLCJlbWFpbCI6ImpjYW5vY0BnbWFpbC5jb20iLCJpYXQiOjE2MjA2NDg2MDEsImV4cCI6MTYyMDY1NTgwMX0.zBMjzcy04a5hJD71eI6FymdAQ8olBgU7V6uX78Mzw-o"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties":{
                "ok":{
                  "type":"boolean",
                  "example":true
                }
              }
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      }
    },
    "/marcas": {
      "get": {
        "tags": ["Marcas"],
        "summary": "Recupera lista de marcas",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "x-token",
            "description": "Token de authenticación",
            "required": true,
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDk4YWI5NzMyMTQyZTA3ZjA2NjU0YjgiLCJuYW1lIjoiSnVsaWFuIENhbm8iLCJlbWFpbCI6ImpjYW5vY0BnbWFpbC5jb20iLCJpYXQiOjE2MjA2NDg2MDEsImV4cCI6MTYyMDY1NTgwMX0.zBMjzcy04a5hJD71eI6FymdAQ8olBgU7V6uX78Mzw-o"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa.",
            "schema": {
              "$ref": "#/definitions/listaMarca"
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      },
      "post":{
        "tags": ["Marcas"],
        "summary": "Registra marca",
        "description": "",
        "produces": ["application/json"],
        "parameters": [          
          {
            "in": "header",
            "name": "x-token",
            "description": "Token de authenticación",
            "required": true,
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDk4YWI5NzMyMTQyZTA3ZjA2NjU0YjgiLCJuYW1lIjoiSnVsaWFuIENhbm8iLCJlbWFpbCI6ImpjYW5vY0BnbWFpbC5jb20iLCJpYXQiOjE2MjA2NDg2MDEsImV4cCI6MTYyMDY1NTgwMX0.zBMjzcy04a5hJD71eI6FymdAQ8olBgU7V6uX78Mzw-o"
          },
          {
            "in": "body",
            "name": "Marca",
            "description": "Objeto de marca",
            "required": true,
            "schema": {
                "$ref": "#/definitions/Marca"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa.",
            "schema": {
              "type": "object",
              "properties":{
                "ok":{
                  "type":"boolean",
                  "example":true
                },
                "msg": {
                  "type":"string",
                  "example":"Registro exitoso"
                }
              }
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      }      
    },
    "/marcas/{id}":{
      "put": {
        "tags": ["Marcas"],
        "summary": "Actualizar una marca",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Marca Id",
            "required": true
          },
          {
            "in": "header",
            "name": "x-token",
            "description": "Token de authenticación",
            "required": true,
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDk4YWI5NzMyMTQyZTA3ZjA2NjU0YjgiLCJuYW1lIjoiSnVsaWFuIENhbm8iLCJlbWFpbCI6ImpjYW5vY0BnbWFpbC5jb20iLCJpYXQiOjE2MjA2NDg2MDEsImV4cCI6MTYyMDY1NTgwMX0.zBMjzcy04a5hJD71eI6FymdAQ8olBgU7V6uX78Mzw-o"
          },
          {
            "in": "body",
            "name": "Marca",
            "description": "Objeto de marca",
            "required": true,
            "schema": {
                "$ref": "#/definitions/Marca"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties":{
                "ok":{
                  "type":"boolean",
                  "example":true
                },
                "msg": {
                  "type":"string",
                  "example":"El registro fue actualizado con exito."
                }
              }
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      },
      "delete": {
        "tags": ["Marcas"],
        "summary": "Elimina una marca",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Marca Id",
            "required": true
          },
          {
            "in": "header",
            "name": "x-token",
            "description": "Token de authenticación",
            "required": true,
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDk4YWI5NzMyMTQyZTA3ZjA2NjU0YjgiLCJuYW1lIjoiSnVsaWFuIENhbm8iLCJlbWFpbCI6ImpjYW5vY0BnbWFpbC5jb20iLCJpYXQiOjE2MjA2NDg2MDEsImV4cCI6MTYyMDY1NTgwMX0.zBMjzcy04a5hJD71eI6FymdAQ8olBgU7V6uX78Mzw-o"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties":{
                "ok":{
                  "type":"boolean",
                  "example":true
                }
              }
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      }
    },
    "/proveedores": {
      "get": {
        "tags": ["Proveedores"],
        "summary": "Recupera lista de proveedores",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "x-token",
            "description": "Token de authenticación",
            "required": true,
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDk4YWI5NzMyMTQyZTA3ZjA2NjU0YjgiLCJuYW1lIjoiSnVsaWFuIENhbm8iLCJlbWFpbCI6ImpjYW5vY0BnbWFpbC5jb20iLCJpYXQiOjE2MjA2NDg2MDEsImV4cCI6MTYyMDY1NTgwMX0.zBMjzcy04a5hJD71eI6FymdAQ8olBgU7V6uX78Mzw-o"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa.",
            "schema": {
              "$ref": "#/definitions/listaProveedor"
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      },
      "post":{
        "tags": ["Proveedores"],
        "summary": "Registra nuevo provedor",
        "description": "",
        "produces": ["application/json"],
        "parameters": [          
          {
            "in": "header",
            "name": "x-token",
            "description": "Token de authenticación",
            "required": true,
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDk4YWI5NzMyMTQyZTA3ZjA2NjU0YjgiLCJuYW1lIjoiSnVsaWFuIENhbm8iLCJlbWFpbCI6ImpjYW5vY0BnbWFpbC5jb20iLCJpYXQiOjE2MjA2NDg2MDEsImV4cCI6MTYyMDY1NTgwMX0.zBMjzcy04a5hJD71eI6FymdAQ8olBgU7V6uX78Mzw-o"
          },
          {
            "in": "body",
            "name": "Proveedor",
            "description": "Objeto de proveedor",
            "required": true,
            "schema": {
                "$ref": "#/definitions/Proveedor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa.",
            "schema": {
              "type": "object",
              "properties":{
                "ok":{
                  "type":"boolean",
                  "example":true
                },
                "msg": {
                  "type":"string",
                  "example":"Registro exitoso"
                }
              }
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      }      
    },
    "/proveedores/{id}":{
      "put": {
        "tags": ["Proveedores"],
        "summary": "Actualizar un proveedor",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Proveedor Id",
            "required": true
          },
          {
            "in": "header",
            "name": "x-token",
            "description": "Token de authenticación",
            "required": true,
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDk4YWI5NzMyMTQyZTA3ZjA2NjU0YjgiLCJuYW1lIjoiSnVsaWFuIENhbm8iLCJlbWFpbCI6ImpjYW5vY0BnbWFpbC5jb20iLCJpYXQiOjE2MjA2NDg2MDEsImV4cCI6MTYyMDY1NTgwMX0.zBMjzcy04a5hJD71eI6FymdAQ8olBgU7V6uX78Mzw-o"
          },
          {
            "in": "body",
            "name": "Proveedor",
            "description": "Objeto de proveedor",
            "required": true,
            "schema": {
                "$ref": "#/definitions/Proveedor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties":{
                "ok":{
                  "type":"boolean",
                  "example":true
                },
                "msg": {
                  "type":"string",
                  "example":"El registro fue actualizado con exito."
                }
              }
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      },
      "delete": {
        "tags": ["Proveedores"],
        "summary": "Elimina un proveedor",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Proveedor Id",
            "required": true
          },
          {
            "in": "header",
            "name": "x-token",
            "description": "Token de authenticación",
            "required": true,
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDk4YWI5NzMyMTQyZTA3ZjA2NjU0YjgiLCJuYW1lIjoiSnVsaWFuIENhbm8iLCJlbWFpbCI6ImpjYW5vY0BnbWFpbC5jb20iLCJpYXQiOjE2MjA2NDg2MDEsImV4cCI6MTYyMDY1NTgwMX0.zBMjzcy04a5hJD71eI6FymdAQ8olBgU7V6uX78Mzw-o"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties":{
                "ok":{
                  "type":"boolean",
                  "example":true
                }
              }
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      }
    }
  },
  "definitions": {
    "body_login": {
      "type": "object",
      "required": [
        "email",
        "password"
      ],
      "properties":{
        "email":{
          "type":"string",
          "example":"jcanocarballo@gmail.com"
        },
        "password":{
          "type":"string",
          "example":"123456"
        }
      }
    },
    "body_register": {
      "type": "object",
      "required": [
        "name",
        "email",
        "password"
      ],
      "properties":{
        "name":{
          "type":"string",
          "example":"Julian Cano Carballo"
        },
        "email":{
          "type":"string",
          "example":"jcanocarballo@gmail.com"
        },
        "password":{
          "type":"string",
          "example":"123456"
        }
      }
    },
    "Usuario": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "listaCategoria": {
      "type": "object",
      "properties":{
        "ok":{
          "type": "boolean",
          "example": true
        },
        "categorias": {
          "type":"array",
          "items":{
            "$ref": "#/definitions/Categoria"
          }
        }
      }
    },
    "Categoria":{
      "type":"object",
      "properties":{
        "id":{
          "type":"integer",
          "example": 1
        },
        "nombre_clave_categoria":{
          "type":"string",
          "example": "Servicio"
        },
        "nombre_completo":{
          "type": "string",
          "example": "Servicio"
        },
        "descripcion": {
          "type":"string",
          "example": "alineacion, balanceo, montajes, etc"
        }
      }
    },
    "listaMarca": {
      "type": "object",
      "properties":{
        "ok":{
          "type": "boolean",
          "example": true
        },
        "marcas": {
          "type":"array",
          "items":{
            "$ref": "#/definitions/Marca"
          }
        }
      }
    },
    "Marca":{
      "type":"object",
      "properties":{
        "id":{
          "type":"integer",
          "example": 1
        },
        "nombre_clave_marca":{
          "type":"string",
          "example": "BOSCH"
        },
        "nombre_completo":{
          "type": "string",
          "example": "BOSCH"
        },
        "descripcion": {
          "type":"string",
          "example": "Auto partes y refacciones eléctricas"
        }
      }
    },
    "listaProveedor": {
      "type": "object",
      "properties":{
        "ok":{
          "type": "boolean",
          "example": true
        },
        "proveedores": {
          "type":"array",
          "items":{
            "$ref": "#/definitions/Proveedor"
          }
        }
      }
    },
    "Proveedor":{
      "type":"object",
      "properties":{
        "id":{
          "type":"integer",
          "example": 1
        },
        "nombre_clave":{
          "type":"string",
          "example": "PROV-GRAL"
        },
        "nombre":{
          "type": "string",
          "example": "Proveedor general"
        },
        "descripcion": {
          "type":"string",
          "example": "Proveedor general de refacciones y autopartes"
        },
        "razon_social": {
          "type":"string",
          "example": "Proveedor general  S.A de C.V"
        },
        "rfc": {
          "type":"string",
          "example": "PG123456789"
        },
        "domicilio": {
          "type":"string",
          "example": "C. 30 #145B x 18 y 20"
        },
        "colonia": {
          "type":"string",
          "example": "Los pinos"
        },
        "municipio": {
          "type":"string",
          "example": "Merida"
        },
        "estado": {
          "type":"string",
          "example": "Yucatan"
        }
      }
    },
    "Response_login":{
      "type":"object",
      "properties":{
        "ok":{
          "type":"boolean",
          "example": true
        },
        "id":{
          "type":"integer",
          "example": 1     
        },
        "name":{
          "type":"string",
          "example":"Julian"
        },
        "email":{
          "type": "string",
          "example":"jcanocarballo@gmail.com"
        },
        "token":{
          "type":"string",
          "example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Ijc4N2I4ZTdjLWI4YzQtMTFlYi1hMjU2LTE2YTk2ZDZlZTNlYiIsIm5hbWUiOiJKdWxpYW4gQ2FubyIsImVtYWlsIjoiamNhbm9jQGdtYWlsLmNvbSIsImlhdCI6MTYyMTQ1ODY5OCwiZXhwIjoxNjIxNDY1ODk4fQ.TN6ixTW7TBpdj9XfI_-dUPcJaA2Xv66EyzeKjZz6sak"
        }
      }
    },
    "400_login":{
      "type":"object",
      "properties":{
        "ok":{
          "type":"boolean",
          "example": false
        },
        "msg":{
          "type":"string",
          "example": "El usuario y/o contraseña no validos"
        }
      }
    },
    "400_renew":{
      "type":"object",
      "properties":{
        "ok":{
          "type":"boolean",
          "example": false
        },
        "msg":{
          "type":"string",
          "example": "Token no válido"
        }
      }
    }
  }
}